requirements.txt
argparse
inquirer
natsort
pyyaml